+++ GamePad/GamePadButtons.lua
@@ -68,7 +68,7 @@ local SetButtonExpanded = [[
       if expandedstate == 4 and type == 2 then return end
       if expandedstate == 5 and type == 3 then return end
 
-      if dclktype > 0 and state == 4 then
+      if (dclktype == 1 and state == 4) or (dclktype == 2  and state ~= 3 and state ~= 5) then
          GamePad:SetAttribute("state-trigger", 4)
          if button == "LeftButton" then
             GamePad:SetAttribute("state-expanded", 1)
@@ -82,6 +82,31 @@ local SetButtonExpanded = [[
                Crosshotbar:RunAttribute("update-expanded")
             end
          end
+      elseif dclktype == 2 then
+         local a = 0
+         if state == 6 or state == 3 then a = 2 end
+         if state == 7 or state == 5 then a = 3 end
+         local b = a - state + 4
+
+         if a == 0 then
+            a = b
+            b = 0
+         end
+
+         local found = true
+         if  a == type then
+            a = 0
+         elseif b == type then
+            b = 0
+         else
+            found = false
+         end
+
+         local newstate = 4
+         if found then
+            newstate = a - b + 4
+         end
+         GamePad:SetAttribute("state-trigger", newstate)
       end
    end
 ]]
@@ -96,6 +121,27 @@ addon.GamePadButtonsMixin = {
 }
 
 local GamePadButtonsMixin = addon.GamePadButtonsMixin
+
+function GamePadButtonsMixin:AddEventHandler(frame)
+   frame:SetAttribute("event-ready", 0)
+   frame:SetAttribute("event-ncount", 0)
+   RegisterAttributeDriver(frame, "event-ready", 1)
+   frame:SetAttributeNoHandler("_onattributechanged", [[
+        local ncount = self:GetAttribute("event-ncount")
+        if name == "event-ready" and value == 1 and ncount > 0 then
+           ncount = ncount - 1
+           self:SetAttribute("event-ncount", ncount)
+           self:SetAttribute("event-ready", 0)
+        end
+    ]])
+end
+
+function GamePadButtonsMixin:RemoveEventHandler(frame)
+   frame:SetAttribute("event-ready", 0)
+   frame:SetAttribute("event-ncount", 0)
+   UnregisterAttributeDriver(frame, "event-ready")
+   frame:SetAttributeNoHandler("_onattributechanged", nil)
+end
 
 function GamePadButtonsMixin:CreatePairButton(ButtonName)
    local Button = CreateFrame("Button", ADDON .. ButtonName .. "ButtonFrame",
@@ -118,70 +164,84 @@ function GamePadButtonsMixin:CreateLeftTriggerButton()
    self.LeftTriggerButton = self:CreatePairButton("LeftTrigger")
    self.LeftTriggerButton:SetAttribute("SetButtonPairState", SetButtonPairState)
    self.LeftTriggerButton:SetAttribute("SetButtonExpanded", SetButtonExpanded)
+   SecureHandlerWrapScript(self.LeftTriggerButton, "OnClick", self.LeftTriggerButton,
+                           [[self:RunAttribute("SetButtonPairState", "LeftButton", down, "trigger")]])
+end
+
+function GamePadButtonsMixin:SetLeftTriggerButtonHandler(type)
+   self:RemoveEventHandler(self.LeftTriggerButton)
+   SecureHandlerUnwrapScript(self.LeftTriggerButton, "OnClick")
+   SecureHandlerUnwrapScript(self.LeftTriggerButton, "OnDoubleClick")
    
-   self.LeftTriggerButton:SetAttribute("dclick-ready", 0)
-   self.LeftTriggerButton:SetAttribute("dclick-ncount", 0)
-   RegisterAttributeDriver(self.LeftTriggerButton, "dclick-ready", 1)
-   self.LeftTriggerButton:SetAttributeNoHandler("_onattributechanged", [[
-        local ncount = self:GetAttribute("dclick-ncount")
-        if name == "dclick-ready" and value == 1 and ncount > 0 then
-           ncount = ncount - 1
-           self:SetAttribute("dclick-ncount", ncount)
-           self:SetAttribute("dclick-ready", 0)
-        end
-    ]])
-   
-   SecureHandlerWrapScript(self.LeftTriggerButton, "OnClick", self.LeftTriggerButton, [[
-        local ready = self:GetAttribute("dclick-ready")
-        if down then
-           if ready == 1 then
-              self:SetAttribute("dclick-ready", 0);
-              self:SetAttribute("dclick-ncount", 1);
-           else
-              self:RunAttribute("SetButtonExpanded", "LeftButton")
-           end
-        else
-           if ready == 0 then
-              self:SetAttribute("dclick-ncount", 1);
-           end
+   if type == 0 then
+      SecureHandlerWrapScript(self.LeftTriggerButton, "OnClick", self.LeftTriggerButton,
+                              [[self:RunAttribute("SetButtonPairState", "LeftButton", down, "trigger")]])
+   elseif type == 1 then
+      self:AddEventHandler(self.LeftTriggerButton)
+      SecureHandlerWrapScript(self.LeftTriggerButton, "OnClick", self.LeftTriggerButton, [[
+      local ready = self:GetAttribute("event-ready")
+         if down then
+            if ready == 1 then
+               self:SetAttribute("event-ready", 0);
+               self:SetAttribute("event-ncount", 1);
+            else
+               self:RunAttribute("SetButtonExpanded", "LeftButton")
+            end
+         else
+            if ready == 0 then
+              self:SetAttribute("event-ncount", 1);
+            end
         end
         self:RunAttribute("SetButtonPairState", "LeftButton", down, "trigger")
    ]])
+   elseif type == 2 then
+      SecureHandlerWrapScript(self.LeftTriggerButton, "OnClick", self.LeftTriggerButton,
+                           [[self:RunAttribute("SetButtonPairState", "LeftButton", down, "trigger")]])
+      SecureHandlerWrapScript(self.LeftTriggerButton, "OnDoubleClick", self.LeftTriggerButton,
+                              [[self:RunAttribute("SetButtonExpanded", "LeftButton")]])
+   end
 end
 
 function GamePadButtonsMixin:CreateRightTriggerButton()
    self.RightTriggerButton = self:CreatePairButton("RightTrigger")
    self.RightTriggerButton:SetAttribute("SetButtonPairState", SetButtonPairState)
    self.RightTriggerButton:SetAttribute("SetButtonExpanded", SetButtonExpanded)
+   SecureHandlerWrapScript(self.RightTriggerButton, "OnClick", self.RightTriggerButton,
+                           [[self:RunAttribute("SetButtonPairState", "RightButton", down, "trigger")]])
+end
+
+function GamePadButtonsMixin:SetRightTriggerButtonHandler(type)
+   self:RemoveEventHandler(self.RightTriggerButton)
+   SecureHandlerUnwrapScript(self.RightTriggerButton, "OnClick")
+   SecureHandlerUnwrapScript(self.RightTriggerButton, "OnDoubleClick")
    
-   self.RightTriggerButton:SetAttribute("dclick-ready", 0)
-   self.RightTriggerButton:SetAttribute("dclick-ncount", 0)
-   RegisterAttributeDriver(self.RightTriggerButton, "dclick-ready", 1)
-   self.RightTriggerButton:SetAttributeNoHandler("_onattributechanged", [[
-        local ncount = self:GetAttribute("dclick-ncount")
-        if name == "dclick-ready" and value == 1 and ncount > 0 then
-           ncount = ncount - 1
-           self:SetAttribute("dclick-ncount", ncount)
-           self:SetAttribute("dclick-ready", 0)
-        end
-    ]])
-   
-   SecureHandlerWrapScript(self.RightTriggerButton, "OnClick", self.RightTriggerButton, [[
-        local ready = self:GetAttribute("dclick-ready")
-        if down then
-           if ready == 1 then
-              self:SetAttribute("dclick-ready", 0);
-              self:SetAttribute("dclick-ncount", 1);
-           else
-              self:RunAttribute("SetButtonExpanded", "RightButton")
-           end
-        else
-           if ready == 0 then
-              self:SetAttribute("dclick-ncount", 1);
-           end
+   if type == 0 then
+      SecureHandlerWrapScript(self.RightTriggerButton, "OnClick", self.RightTriggerButton,
+                              [[self:RunAttribute("SetButtonPairState", "RightButton", down, "trigger")]])
+   elseif type == 1 then
+      self:AddEventHandler(self.RightTriggerButton)
+      SecureHandlerWrapScript(self.RightTriggerButton, "OnClick", self.RightTriggerButton, [[
+      local ready = self:GetAttribute("event-ready")
+         if down then
+            if ready == 1 then
+               self:SetAttribute("event-ready", 0);
+               self:SetAttribute("event-ncount", 1);
+            else
+               self:RunAttribute("SetButtonExpanded", "RightButton")
+            end
+         else
+            if ready == 0 then
+              self:SetAttribute("event-ncount", 1);
+            end
         end
         self:RunAttribute("SetButtonPairState", "RightButton", down, "trigger")
    ]])
+   elseif type == 2 then
+      SecureHandlerWrapScript(self.RightTriggerButton, "OnClick", self.RightTriggerButton,
+                           [[self:RunAttribute("SetButtonPairState", "RightButton", down, "trigger")]])
+      SecureHandlerWrapScript(self.RightTriggerButton, "OnDoubleClick", self.RightTriggerButton,
+                              [[self:RunAttribute("SetButtonExpanded", "RightButton")]])
+   end
 end
 
 function GamePadButtonsMixin:CreateLeftShoulderButton()